organize views by profile type they serve
